% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.r
\docType{class}
\name{WoMNetwork-class}
\alias{WoMNetwork-class}
\title{Class \code{WoMNetwork}}
\description{
Class \code{WoMNetwork}
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("WoMNetwork", ...)}. This S4 class describes \code{WoMNetwork} objects.
}

\examples{

network <- new("WoMNetwork", size = 1000, avgConnections = 5)
print(network)

}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}

Thomas Woehner \email{Thomas.Woehner@eah-jena.de}

Ralf Peters \email{ralf.peters@wiwi.uni-halle.de}
}
\keyword{classes}
