% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.r
\docType{methods}
\name{show,WoMNetwork-method}
\alias{show,WoMNetwork-method}
\title{Shows a \code{WoMNetwork} object}
\usage{
\S4method{show}{WoMNetwork}(object)
}
\arguments{
\item{object}{An instance of the \code{WoMNetwork}-class}
}
\description{
Shows a \code{WoMNetwork} object
}
\section{Methods}{
 \describe{
\item{list("signature(object = \"WoMNetwork\")")}{Shows an \code{WoMNetwork} object.}}
}

\author{
Michael Scholz \email{michael.scholz@th-deg.de}

Thomas Woehner \email{Thomas.Woehner@eah-jena.de}

Ralf Peters \email{ralf.peters@wiwi.uni-halle.de}
}
\keyword{methods}
