% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordOfMouth-package.r
\docType{package}
\name{WordOfMouth-package}
\alias{WordOfMouth-package}
\alias{WordOfMouth}
\title{Estimates Economic Variables for Word-of-Mouth-Campaigns}
\description{
This packages provides classes, methods and functions for modeling Word-of-Mouth-campaigns.
General model assumptions are: 
\itemize{
\item monopoly market
\item no variable costs
\item network is the mean-field case of percolation
\item only those persons who bought a product will forward information about it
}
}
\details{
\tabular{ll}{ Package: \tab WordOfMouth\cr Type: \tab Package\cr Version: 
\tab 1.0.0\cr Date: \tab 2020-03-27\cr License: \tab GPL-3\cr Depends: \tab
R (>= 3.0.1), methods\cr }
}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}

Thomas Woehner \email{Thomas.Woehner@eah-jena.de}

Ralf Peters \email{ralf.peters@wiwi.uni-halle.de}
}
\keyword{Word-of-Mouth}
